requirements.txt
aiogram
rich
pytube
requests
pyyaml